requirement.txt
MarkupSafe                      2.1.5
PyQt6                           6.7.1
PyQt6-Qt6                       6.7.2
PyQt6-WebEngine                 6.7.0
PyQt6-WebEngine-Qt6             6.7.2
PyQt6-WebEngineSubwheel-Qt6     6.7.2
PyQt6-sip                       13.8.0
branca                          0.7.2
certifi                         2024.7.4
charset-normalizer              3.3.2
folium                          0.17.0
geographiclib                   2.0
geopandas                       1.0.1
geopy                           2.4.1
idna                            3.7
jinja2                          3.1.4
joblib                          1.4.2
mysql-connector-python          9.0.0
networkx                        3.2.1
numpy                           2.0.1
osmnx                           1.9.3
packaging                       24.1
pandas                          2.2.2
pip                             23.2.1
pyogrio                         0.9.0
pyproj                          3.6.1
python-dateutil                 2.9.0.post0
pytz                            2024.1
requests                        2.32.3
scikit-learn                    1.5.1
scipy                           1.13.1
setuptools                      68.2.0
shapely                         2.0.5
six                             1.16.0
threadpoolctl                   3.5.0
tzdata                          2024.1
urllib3                         2.2.2
wheel                           0.41.2
xyzservices                     2024.6.0